Moved renderable nav categories into redux store
