+++ src/reducers/navigationReducer.js
@@ -25,6 +25,11 @@ export default function navigationReducer(state: NavigationState = initialState,
         ...state,
         isFetching: false,
       };
+    case "SET_RENDERABLE_NAVIGATION_CATEGORIES":
+      return {
+        ...state,
+        renderableNavigationCategories: action.categories,
+      };
     default:
       return state;
   }
