+++ src/store/configureStore.js
@@ -6,6 +6,7 @@ import { persistStore, persistCombineReducers } from "redux-persist";
 import reducers from "../reducers";
 import offlineDataMiddleware from "../middleware/offlineDataMiddleware";
 import audioMiddleware from "../middleware/audioMiddleware";
+import navigationMiddleware from "../middleware/navigationMiddleware";
 
 const config = {
   key: "hbgRoot",
@@ -20,6 +21,7 @@ const reducer = persistCombineReducers(config, reducers);
 const middlewares = [
   offlineDataMiddleware,
   audioMiddleware,
+  navigationMiddleware,
   thunk,
 ];
